Ignore port when comparing host regexes
Switches the matching function to use URL.hostname instead of
URL.host. The former doesn't include the port, which allows
matching the domain regardless of whether a port is included
in the request. This is rarely needed since most Google domains
just use standard HTTP/HTTPS ports, but it can be useful in
rare circumstances.